Add basic JSON support and demo it
First we add the `json()` method to `ServerResponse`.
It can take an arbitrary `Codable` object and send
it to the client as JSON.

Further we add the `Todo` struct, as a Codable, as
well as a few demo instances.

Finally we send our in-memory todo database to the
client using a middleware:

    app.get("/todomvc") { _, res, _ in
      try res.json(todos) // send JSON to the browser
    }